rename event to events because of documentation issues
